backend/requirements.txt
annotated-types==0.7.0
anyio==4.6.1
bcrypt==4.2.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.1
dnspython==2.7.0
docker==7.1.0
email_validator==2.2.0
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
fsspec==2024.9.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.1
httpx==0.27.2
huggingface-hub==0.25.2
idna==3.10
Jinja2==3.1.4
# jwt==1.3.1
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==3.0.1
mdurl==0.1.2
mpmath==1.3.0
networkx==3.4.1
numpy==2.1.2
packaging==24.1
passlib==1.7.4
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pymongo==4.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
PyYAML==6.0.2
regex==2024.9.11
requests==2.32.3
rich==13.9.2
safetensors==0.4.5
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.39.2
sympy==1.13.3
tokenizers==0.20.1
torch==2.4.1
tqdm==4.66.5
transformers==4.45.2
typer==0.12.5
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.31.1
uvloop==0.20.0
watchfiles==0.24.0
websockets==13.1
PyJWT==2.9.0
